Update TemporalMemoryTests_2.cs
In this unit test,a new instance of the TemporalMemory class has been created and initialize it with a new instance of Connections and some default parameters.
Then set the column dimensions to be a 2D array of size 62x62,and the number of cells per column to be 30.After that we applied these parameters to the connections object and initialize the temporal memory with the connections.
Next, we get the actual number of columns from the connections object and compare it to the expected number of columns (which is calculated as the product of the dimensions).
We use the Assert.AreEqual() method to make sure that the actual number of columns matches the expected number of columns.